rush00

+++ rush00/panier.php
@@ -12,15 +12,29 @@
 		return false;
 	}
 
+	function notExistindb($data, $name) {
+		$i = 0;
+		foreach ($data as $k => $v) {
+			foreach ($v as $k2 => $v2) {
+				if ($v2['name'] === $name) {
+					$i = 1;
+				}
+			}
+		}
+		return ($i === 1) ? true : false;
+	}
+
 	function display_list_admin($tab) {
 		foreach ($tab as $product) {
 			echo '<ul class="admin-product">';
 				echo '<li><span>iD: </span>' . $product['id'] . '</li>';
 				echo '<li><span>Name: </span>' . $product['name'] . '</li>';
 				echo '<li><span>Brand: </span>' . $product['brand'] . '</li>';
-						foreach ($product['categories'] as $categorie) {
-							echo '<li>' . ucfirst($categorie) . '&nbsp;&nbsp;</li>';
-						}
+				if ($products['categories']) {
+					foreach ($product['categories'] as $categorie) {
+						echo '<li>' . ucfirst($categorie) . '&nbsp;&nbsp;</li>';
+					}
+				}
 				echo '<li><span>Price: </span>' . $product['price'] . ' €</li>';
 				echo '<li><a href="?action=del?id=' . $product['id'] . '"><i class="fa fa-trash"></i></a></li> <br />';
 				echo '<button class="delete" type="submit"><a id="adel" href="panier.php?del=' . $product['name'] . '">Supprimer</a></button>';
@@ -48,8 +62,10 @@ if ($_GET['name']) {
 			if (already_exists($panier, $_GET['name'])) {
 				echo "<h2 id='already'>Vous avez déjà le chat " . $_GET['name'] . "</h2>";
 			} else {
-				$panier[] = $tab;
-				file_put_contents('private/panier',serialize($panier));
+				if (notExistIndb($data, $_GET['name'])) {
+					$panier[] = $tab;
+					file_put_contents('private/panier',serialize($panier));
+				}
 			}
 		}
 }
@@ -77,6 +93,7 @@ if ($_GET['del']) {
 			if ($panier) {
 				display_list_admin($panier);
 			}
+			echo "<button id='butpan' type='submit'><a id='apan' href='thx.php'>Valider le panier</a></button>";
 		 ?>
 
  		<?php include_once('views/footer.html'); ?>
